% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docs_test}
\alias{docs_test}
\title{Document dataset}
\format{a dataset with 1000 documents}
\usage{
docs_test
}
\description{
A dataset of 1000 documents extracted from WoS repository
}
\keyword{datasets}
