% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{lda_arrange_docs}
\alias{lda_arrange_docs}
\title{Arrange the topic distribution per document}
\usage{
lda_arrange_docs(lda, n_ranks = 1, n_docs = 0)
}
\arguments{
\item{lda}{An LDA object}

\item{n_ranks}{Number of topic ranks (<=k)}

\item{n_docs}{Number of returned documents per highest ranking topic. By
default reurns all documents}
}
\value{
A table of n_docs documents per main topic, each associated with
  n_ranks topic and its gamma
}
\description{
Create a table were each row lists a document and its topic distribution
(gamma) in a descending order
}
