% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{lda_expose_docs}
\alias{lda_expose_docs}
\title{Expose document info by their lead topic}
\usage{
lda_expose_docs(lda, n_docs = 0, ranks = 1, topic_names = NULL,
  metadata = NULL)
}
\arguments{
\item{lda}{An LDA object}

\item{n_docs}{Number of returned documents per topic. If 0 returns all
documents}

\item{ranks}{Ranks of topics to return}

\item{topic_names}{A 2 column tibble - topic (key) and topic name}

\item{metadata}{A data frame of document metadata. Key column - document}
}
\value{
A table of n_docs documents per topic, each associated with the
  requested info
}
\description{
For each document creates a table of documents associated with the highest
ranking topic and its gamma value. Optionally, The function extends the info
with the document metadata and lower ranking topics. Topic numbers can be
replaced with topic names
}
