% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_docs.R
\name{lda_list_docs}
\alias{lda_list_docs}
\title{List documents by their lead topic}
\usage{
lda_list_docs(lda, n_ranks = 1, n_docs = 0, topics = NULL)
}
\arguments{
\item{lda}{LDA object}

\item{n_ranks}{Number of topic ranks to display (<=k). Default 1}

\item{n_docs}{Number of returned documents per lead topic. Default all}

\item{topics}{A list of topic numbers to be displayed}
}
\value{
A list of n_docs documents per lead topic, each associated with
  n_ranks topic and their gamma
}
\description{
A topic modeling utility. List documents by their most representative topic
based on the gamma table of the LDA_ object resulting from
`topicmodels::LDA()`. Optionally, lower ranking topics per document are also
displayed
}
