% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_topics.R
\name{lda_table_topics}
\alias{lda_table_topics}
\title{Spread the term distribution over topics by topic columns}
\usage{
lda_table_topics(lda, n_terms = 20, topics = NULL, beta = TRUE)
}
\arguments{
\item{lda}{LDA_ object}

\item{n_terms}{Number of highest ranking terms per topic to display}

\item{topics}{A list of topic numbers to display}

\item{beta}{retrun beta values? default all}
}
\value{
a table that spreads for each topic a column of terms and their
  associated beta
}
\description{
A topic modeling utility. Spreads the beta table of the LDA_
  object resulting from `topicmodels::LDA()` into the higest ranking terms
  and their beta values in columns per topic. The number of requested terms
  and topics can be specified as well as whether to display the beta's
}
