% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LDA_test}
\alias{LDA_test}
\title{LDA object for testing}
\format{A LDA_Gibbs object of 10 topics, 1814 documents and 8008 terms}
\usage{
LDA_test
}
\description{
This LDA object is the result of topic modeling the copus in the VR
Rehabilitation project
}
\keyword{datasets}
