% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_topics.R
\name{lda_trim_beta}
\alias{lda_trim_beta}
\title{Trim LDA beta table by terms and/or topics}
\usage{
lda_trim_beta(lda, n_terms = 20, topics = NULL)
}
\arguments{
\item{lda}{LDA object}

\item{n_terms}{Number of highest ranking terms per topic to display. Default 20}

\item{topics}{A list of topic numbers to return}
}
\value{
a tibble of highest ranking terms per topic and their beta values
}
\description{
A topic modeling utility. Trims the beta table of the LDA
  object resulting from `topicmodels::LDA()` by the requested number of
  highest ranking terms and / or topics
}
