% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_stopwords}
\alias{my_stopwords}
\title{Custom stopwords}
\format{43 rows of one 'word' column}
\usage{
my_stopwords
}
\description{
A tibble with one column of custom stopwords
}
\keyword{datasets}
