% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_url.R
\name{scrape_text}
\alias{scrape_text}
\title{Scrape text from URL}
\usage{
scrape_text(url, tag)
}
\arguments{
\item{url}{the required URL}

\item{tag}{CSS node tag of required text}
}
\value{
A text object that combines all segments of the node tag
}
\description{
Scrape text of a specific node tag in a given URL. All text
  segments are combine to one text object
}
