% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_text.R
\name{tm_make_corpus}
\alias{tm_make_corpus}
\title{Create a clean, stemmed and tidy corpus}
\usage{
tm_make_corpus(documents, custom_sw = NULL, stem2readable = TRUE)
}
\arguments{
\item{documents}{object. Must contain a 'text' column}

\item{custom}{stopwords object. Must contain a 'word' column. Optional}

\item{stem}{to readable flag. defaults to TRUE}
}
\value{
a corpus that includes the original word and the stemmed word in the 'stem' column
}
\description{
A text cleanup utility. takes in a dataset of documents,
  cleans it, removes stopwords and performs Porter stemming using the SnowballC
  package. It outputs a tidy corpus with stemmed words. Optionally is takes a
  custom stopwords list and optionally transforms the stemmed words to a readable form
}
