% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_text.R
\name{tm_stem2readable}
\alias{tm_stem2readable}
\title{Convert stemmed words to a readable format}
\usage{
tm_stem2readable(corpus)
}
\arguments{
\item{corpus}{object in a tidy format containing at least 'word' 'stem' columns}
}
\value{
the corpus where the 'stem' column is replaces with a readable word
}
\description{
A text cleanup utility. It converts the words that have been
  stemmed to a readable format by selecting the most frequent word among those
  transformed to the same root.
}
