% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmutilsr.R
\docType{package}
\name{tmutilsr-package}
\alias{tmutilsr}
\alias{tmutilsr-package}
\title{tmutilsr: Text mining utilities}
\description{
tmutilsr extends basic R text mining packages with commonly used processing
functions
}
\details{
Topic modeling utilities (base package `topicmodels`):

\itemize{
\item `lda_trim_beta` - trims LDA beta table by terms and/or topics
\item `lda_table_topics` - spreads the term distribution over topics by topic
\item `lda_list_docs` - lists documents by their lead topic
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/doritge/tmutilsr}
  \item Report bugs at \url{http://github.com/doritge/tmutilsr/issues}
}

}
\author{
\strong{Maintainer}: Dorit Geifman \email{dorit.geifman@gmail.com}

}
\keyword{internal}
