% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_topics.R
\name{lda_trim_beta}
\alias{lda_trim_beta}
\title{Trim LDA beta table by terms and/or topics}
\usage{
lda_trim_beta(lda, n_terms = 20, topics = NULL)
}
\arguments{
\item{lda}{LDA_ object}

\item{n_terms}{Number of highest ranking terms per topic to display}

\item{topics}{A list of topic numbers to display}
}
\value{
a data frame of highest ranking terms per topic and their beta values
}
\description{
A topic modeling utility. Trims the beta table of the LDA_
  object resulting from `topicmodels::LDA()` by the requested number of
  highest ranking terms and / or topics
}
